requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1 